Refact/#34 complain,complain feedback,opinion refact with login

+++ backend/src/main/java/com/ohammer/apartner/domain/complaint/service/ComplaintFeedbackService.java
@@ -6,7 +6,7 @@ import com.ohammer.apartner.domain.complaint.entity.Complaint;
 import com.ohammer.apartner.domain.complaint.entity.ComplaintFeedback;
 import com.ohammer.apartner.domain.complaint.repository.ComplaintFeedbackRepository;
 import com.ohammer.apartner.domain.user.entity.User;
-import com.ohammer.apartner.domain.user.repository.UserRepository;
+import com.ohammer.apartner.security.utils.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -18,7 +18,6 @@ import java.util.stream.Collectors;
 public class ComplaintFeedbackService {
     private final ComplaintFeedbackRepository complaintFeedbackRepository;
     private final ComplaintService complaintService;
-    private final UserRepository userRepository;
 
     // Read
     public List<AllComplaintFeedbackResponseDto> findComplaintFeedbackByComplaintId(Long complaintId) {
@@ -41,20 +40,27 @@ public class ComplaintFeedbackService {
         Complaint complaint = complaintService.findById(complaintFeedbackRequestDto.getComplaintId());
 
         // 로그인 로직에 따라 변경 필요
-//        User user = userRepository.findById(userId);
+        User user = SecurityUtil.getCurrentUser();
 
         ComplaintFeedback complaintFeedback = ComplaintFeedback.builder()
                 .content(complaintFeedbackRequestDto.getContent())
                 .complaint(complaint)
-//                .user(user)
+                .user(user)
                 .build();
 
         return complaintFeedbackRepository.save(complaintFeedback);
     }
 
     // Update
-    public ComplaintFeedback updateComplaintFeedback(Long feedbackId, CreateComplaintFeedbackRequestDto complaintFeedbackRequestDto){
+    public ComplaintFeedback updateComplaintFeedback(Long feedbackId, CreateComplaintFeedbackRequestDto complaintFeedbackRequestDto) throws Exception {
         ComplaintFeedback complaintFeedback = findByFeedbackId(feedbackId);
+
+        User user = SecurityUtil.getCurrentUser();
+
+        if(!complaintFeedback.getUser().getId().equals(user.getId())){
+            throw new Exception("유저의 Id가 다릅니다.");
+        }
+
         complaintFeedback.setContent(complaintFeedbackRequestDto.getContent());
         return complaintFeedbackRepository.save(complaintFeedback);
     }
@@ -64,15 +70,15 @@ public class ComplaintFeedbackService {
         return complaintFeedbackRepository.findById(feedbackId).orElse(null);
     }
 
-    public void deleteComplainFeedback(Long feedbackId) {
+    public void deleteComplainFeedback(Long feedbackId) throws Exception {
 
-        Long findId = complaintFeedbackRepository.findUserIdById(feedbackId);
+        Long userId = complaintFeedbackRepository.findUserIdById(feedbackId);
 
         // 작성자와 지우는 시람비교 혹은 관리자인지
-//        userId =
-//        if(findId.equals(userid)){
-//
-//        }
+        User user = SecurityUtil.getCurrentUser();
+        if(!user.getId().equals(userId)){
+            throw new Exception("유저의 ID가 다릅니다");
+        }
         complaintFeedbackRepository.deleteById(feedbackId);
     }
 }
